+++ client/src/App.jsx
@@ -10,8 +10,8 @@ import SupplementsComponent from "./components/Supplements";
 import AllProductsComponent from "./components/AllProducts";
 import ProductOverview from "./components/ProductOverview/ProductOverview";
 import Checkout from "./components/Checkout";
-import OrderConfirmation from "./components/OrderConfirmation";
-import OrderFulfillment from "./components/OrderFulfillment";
+import OrderConfirmation from "./components/Order/OrderConfirmation";
+import OrderFulfillment from "./components/Order/OrderFulfillment";
 import CheckoutButton from "./components/CheckoutButton";
 import EquipmentComponent from "./components/Equipment";
 import ApparelComponent from "./components/Apparel";
